+++ docs/changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Add warning [when Android Jetifier is enabled.](https://github.com/runningcode/gradle-doctor/pull/118)
+* Print tasks with [negative avoidance savings from the cache.](https://github.com/runningcode/gradle-doctor/pull/117) [Fixes #86](https://github.com/runningcode/gradle-doctor/issues/86)
 
 ## 0.5.1
 * Fix bugs in JAVA_HOME check [#110](https://github.com/runningcode/gradle-doctor/pull/110). Thanks [ZacSweers](https://github.com/ZacSweers)